add cli argument to clear db on server start: "fresh=true"; add goland script configs for running e2e and unit tests as well as javascript linting
