+++ .idea/runConfigurations/e2e__browser_.xml
@@ -1,0 +1,12 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="e2e (browser)" type="js.build_tools.npm">
+    <package-json value="$PROJECT_DIR$/e2e/package.json" />
+    <command value="run" />
+    <scripts>
+      <script value="open" />
+    </scripts>
+    <node-interpreter value="project" />
+    <envs />
+    <method v="2" />
+  </configuration>
+</component>
