+++ .idea/runConfigurations/start.xml
@@ -1,8 +1,8 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="go build github.com/nrc-no/core/api/cmd/server" type="GoApplicationRunConfiguration" factoryName="Go Application" nameIsGenerated="true">
+  <configuration default="false" name="start" type="GoApplicationRunConfiguration" factoryName="Go Application">
     <module name="core" />
     <working_directory value="$PROJECT_DIR$/api" />
-    <parameters value="--mongo-servers &quot;mongodb://localhost:27017,mongodb://localhost:27018,mongodb://localhost:27019&quot; --mongo-database dev --bind-address &quot;127.0.0.1&quot; --bind-port 8000 --oidc-issuer-url &quot;http://localhost:8080/auth/realms/Core&quot; --oidc-client-id api --oidc-client-secret a9a1b0f8-3920-4a6a-9c5f-2ed9e49e833e" />
+    <parameters value="--mongo-database=core --mongo-username=root --mongo-password=example --environment=Development --fresh=true" />
     <EXTENSION ID="net.ashald.envfile">
       <option name="IS_ENABLED" value="false" />
       <option name="IS_SUBST" value="false" />
@@ -14,9 +14,9 @@
       </ENTRIES>
     </EXTENSION>
     <kind value="PACKAGE" />
-    <package value="github.com/nrc-no/core/api/cmd/server" />
+    <package value="github.com/nrc-no/core/cmd" />
     <directory value="$PROJECT_DIR$" />
-    <filePath value="$PROJECT_DIR$/api/pkg/server2/cmd/server/server.go" />
+    <filePath value="$PROJECT_DIR$/cmd/main.go" />
     <method v="2" />
   </configuration>
 </component>
