--- .run/openapi.yaml.run.xml
@@ -1,18 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="openapi.yaml" type="SwaggerCodegen">
-    <option name="activeConfigurationType" value="FILE" />
-    <option name="configurationPath" value="" />
-    <option name="customJreEnabled" value="false" />
-    <option name="customTemplatesPath" value="" />
-    <option name="enableDebugLogging" value="false" />
-    <option name="generateToPath" value="$PROJECT_DIR$/gen" />
-    <option name="generationParameters">
-      <list />
-    </option>
-    <option name="generatorPath" value="https://repo1.maven.org/maven2/org/openapitools/openapi-generator-cli/5.0.0/openapi-generator-cli-5.0.0.jar" />
-    <option name="jdkPath" value="Default" />
-    <option name="language" value="go-gin-server" />
-    <option name="specificationPath" value="$PROJECT_DIR$/libs/openapi/openapi.yaml" />
-    <method v="2" />
-  </configuration>
-</component>
