--- api/.run/start.run.xml
@@ -1,22 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="start" type="GoApplicationRunConfiguration" factoryName="Go Application">
-    <module name="core" />
-    <working_directory value="$PROJECT_DIR$/api" />
-    <parameters value="--mongo-database=core --mongo-username=root --mongo-password=example --environment Development" />
-    <EXTENSION ID="net.ashald.envfile">
-      <option name="IS_ENABLED" value="false" />
-      <option name="IS_SUBST" value="false" />
-      <option name="IS_PATH_MACRO_SUPPORTED" value="false" />
-      <option name="IS_IGNORE_MISSING_FILES" value="false" />
-      <option name="IS_ENABLE_EXPERIMENTAL_INTEGRATIONS" value="false" />
-      <ENTRIES>
-        <ENTRY IS_ENABLED="true" PARSER="runconfig" />
-      </ENTRIES>
-    </EXTENSION>
-    <kind value="PACKAGE" />
-    <package value="github.com/nrc-no/core/cmd" />
-    <directory value="$PROJECT_DIR$" />
-    <filePath value="$PROJECT_DIR$/cmd/main.go" />
-    <method v="2" />
-  </configuration>
-</component>
