+++ e2e/cypress/support/commands.js
@@ -42,16 +42,9 @@
 //     });
 // });
 
-
-Cypress.Commands.add('resetDB', () => {
-  cy.request('POST', 'localhost:9000/seed', {
-    action: 'RESET'
-  });
-});
-
 Cypress.Commands.add('login', (email) => {
-  cy.intercept(
-    { url: '*' },
-    (req) => (req.headers['X-E2E-Authenticated-User-Email'] = `${email}`)
-  );
+    cy.intercept(
+        { url: '*' },
+        (req) => (req.headers['X-E2E-Authenticated-User-Email'] = `${email}`)
+    );
 });
