+++ e2e/cypress/support/index.js
@@ -20,5 +20,4 @@ import './commands';
 // require('./commands')
 
 // call our custom login with access token before each test
-before(() => cy.resetDB());
 beforeEach(() => cy.login('birdie.tifawt@email.com'));
